pkg/metrics: rename pkg policystatemetrics to policymetrics
This will now be used not only for the state but also for the kernel
memory so I'm making the name more generic.
