+++ pkg/metricsconfig/healthmetrics.go
@@ -16,7 +16,7 @@ import (
 	"github.com/cilium/tetragon/pkg/metrics/kprobemetrics"
 	"github.com/cilium/tetragon/pkg/metrics/opcodemetrics"
 	"github.com/cilium/tetragon/pkg/metrics/policyfiltermetrics"
-	"github.com/cilium/tetragon/pkg/metrics/policystatemetrics"
+	"github.com/cilium/tetragon/pkg/metrics/policymetrics"
 	"github.com/cilium/tetragon/pkg/metrics/watchermetrics"
 	"github.com/cilium/tetragon/pkg/observer"
 	"github.com/cilium/tetragon/pkg/process"
@@ -87,8 +87,8 @@ func registerHealthMetrics(group metrics.Group) {
 	// kprobe metrics
 	kprobemetrics.RegisterMetrics(group)
 	group.ExtendInitForDocs(kprobemetrics.InitMetricsForDocs)
-	// policy state metrics
-	group.MustRegister(policystatemetrics.NewPolicyStateCollector())
+	// policy metrics
+	group.MustRegister(policymetrics.NewPolicyCollector())
 	// gRPC metrics
 	group.MustRegister(grpcmetrics.NewServerMetrics())
 	// missed metris
